+++ geometry/BUILD.bazel
@@ -85,7 +85,7 @@ drake_cc_library(
         "//geometry/proximity:distance_to_point_callback",
         "//geometry/proximity:distance_to_shape_callback",
         "//geometry/proximity:find_collision_candidates_callback",
-        "//geometry/proximity:hydroelastic_callback",
+        "//geometry/proximity:hydroelastic_calculator",
         "//geometry/proximity:obj_to_surface_mesh",
         "//geometry/proximity:penetration_as_point_pair_callback",
         "@fcl_internal//:fcl",
@@ -852,7 +852,7 @@ drake_cc_googletest(
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:expect_no_throw",
         "//common/test_utilities:expect_throws_message",
-        "//geometry/proximity:hydroelastic_callback",
+        "//geometry/proximity:hydroelastic_calculator",
         "//geometry/proximity:make_sphere_mesh",
         "//math",
         "@fcl_internal//:fcl",
