+++ geometry/proximity/BUILD.bazel
@@ -359,9 +359,9 @@ drake_cc_library(
 )
 
 drake_cc_library(
-    name = "hydroelastic_callback",
-    srcs = ["hydroelastic_callback.cc"],
-    hdrs = ["hydroelastic_callback.h"],
+    name = "hydroelastic_calculator",
+    srcs = ["hydroelastic_calculator.cc"],
+    hdrs = ["hydroelastic_calculator.h"],
     internal = True,
     visibility = [
         "//geometry:__pkg__",
@@ -1247,9 +1247,9 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "hydroelastic_callback_test",
+    name = "hydroelastic_calculator_test",
     deps = [
-        ":hydroelastic_callback",
+        ":hydroelastic_calculator",
         "//common/test_utilities",
         "//math:autodiff",
     ],
