+++ geometry/proximity/hydroelastic_calculator.h
